--- a/build/dependencies.props
+++ b/build/dependencies.props
@@ -3,10 +3,10 @@
     <BenchmarkDotNetPackageVersion>0.11.3</BenchmarkDotNetPackageVersion>
     <ConfigureAwaitCheckerAnalyzerPackageVersion>3.0.0</ConfigureAwaitCheckerAnalyzerPackageVersion>
     <GoogleProtobufPackageVersion>3.7.0</GoogleProtobufPackageVersion>
-    <GrpcCorePackageVersion>1.20.0-pre3</GrpcCorePackageVersion>
-    <GrpcCoreApiPackageVersion>1.20.0-pre3</GrpcCoreApiPackageVersion>
-    <GrpcReflectionPackageVersion>1.20.0-pre3</GrpcReflectionPackageVersion>
-    <GrpcToolsPackageVersion>1.20.0-pre3</GrpcToolsPackageVersion>
+    <GrpcCorePackageVersion>1.21.0-pre1</GrpcCorePackageVersion>
+    <GrpcCoreApiPackageVersion>1.21.0-pre1</GrpcCoreApiPackageVersion>
+    <GrpcReflectionPackageVersion>1.21.0-pre1</GrpcReflectionPackageVersion>
+    <GrpcToolsPackageVersion>1.21.0-pre1</GrpcToolsPackageVersion>
     <MicrosoftAspNetCoreAuthenticationJwtBearerPackageVersion>3.0.0-preview-19075-0444</MicrosoftAspNetCoreAuthenticationJwtBearerPackageVersion>
     <MicrosoftAspNetCoreTestHostPackageVersion>3.0.0-preview6-19260-04</MicrosoftAspNetCoreTestHostPackageVersion>
     <MicrosoftExtensionsLoggingTestingPackageVersion>2.0.0-*</MicrosoftExtensionsLoggingTestingPackageVersion>