--- a/build_and_test.sh
+++ b/build_and_test.sh
@@ -18,7 +18,7 @@ set -ex
 source activate.sh
 
 # Required when using nightly builds of gRPC packages
-./build/get-grpc.sh
+# ./build/get-grpc.sh
 
 dotnet build