--- a/perf/Grpc.AspNetCore.Microbenchmarks/Internal/TestPipeWriter.cs
+++ b/perf/Grpc.AspNetCore.Microbenchmarks/Internal/TestPipeWriter.cs
@@ -17,9 +17,7 @@
 #endregion
 
 using System;
-using System.Collections.Generic;
 using System.IO.Pipelines;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;