--- a/src/Grpc.AspNetCore.Server/Internal/CallHandlers/ClientStreamingServerCallHandler.cs
+++ b/src/Grpc.AspNetCore.Server/Internal/CallHandlers/ClientStreamingServerCallHandler.cs
@@ -19,7 +19,6 @@
 using System;
 using System.Threading.Tasks;
 using Grpc.Core;
-using Grpc.Core.Interceptors;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -32,11 +31,26 @@ namespace Grpc.AspNetCore.Server.Internal.CallHandlers
         where TService : class
     {
         private readonly ClientStreamingServerMethod<TService, TRequest, TResponse> _invoker;
-        private static readonly Func<Interceptor, ClientStreamingServerMethod<TRequest, TResponse>, ClientStreamingServerMethod<TRequest, TResponse>> BuildInvoker = (interceptor, next) =>
+        private static readonly Func<InterceptorRegistration, ClientStreamingServerMethod<TRequest, TResponse>, ClientStreamingServerMethod<TRequest, TResponse>> BuildInvoker = (interceptorRegistration, next) =>
         {
-            return (requestStream, context) =>
+            return async (requestStream, context) =>
             {
-                return interceptor.ClientStreamingServerHandler(requestStream, context, next);
+                var interceptorActivator = (IGrpcInterceptorActivator)context.GetHttpContext().RequestServices.GetRequiredService(interceptorRegistration.ActivatorType);
+                var interceptorInstance = interceptorActivator.Create(interceptorRegistration.Args);
+
+                if (interceptorInstance == null)
+                {
+                    throw new InvalidOperationException($"Could not construct Interceptor instance for type {interceptorRegistration.Type.FullName}");
+                }
+
+                try
+                {
+                    return await interceptorInstance.ClientStreamingServerHandler(requestStream, context, next);
+                }
+                finally
+                {
+                    interceptorActivator.Release(interceptorInstance);
+                }
             };
         };
 
@@ -58,8 +72,8 @@ namespace Grpc.AspNetCore.Server.Internal.CallHandlers
 
             var activator = httpContext.RequestServices.GetRequiredService<IGrpcServiceActivator<TService>>();
             TService service = null;
-
             TResponse response = null;
+
             try
             {
                 serverCallContext.Initialize();
@@ -83,11 +97,10 @@ namespace Grpc.AspNetCore.Server.Internal.CallHandlers
                         resolvedContext);
                     };
 
-                    // The list is reversed during construction so the first interceptor is invoked first
+                    // The list is reversed during construction so the first interceptor is built last and invoked first
                     for (var i = ServiceOptions.Interceptors.Count - 1; i >= 0; i--)
                     {
-                        var interceptor = CreateInterceptor(ServiceOptions.Interceptors[i], httpContext.RequestServices);
-                        resolvedInvoker = BuildInvoker(interceptor, resolvedInvoker);
+                        resolvedInvoker = BuildInvoker(ServiceOptions.Interceptors[i], resolvedInvoker);
                     }
 
                     response = await resolvedInvoker(