--- a/src/Grpc.AspNetCore.Server/Internal/DefaultGrpcInterceptorActivator.cs
+++ b/src/Grpc.AspNetCore.Server/Internal/DefaultGrpcInterceptorActivator.cs
@@ -18,7 +18,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using Grpc.Core.Interceptors;
 using Microsoft.Extensions.DependencyInjection;