--- a/src/Grpc.AspNetCore.Server/Internal/GrpcServiceBinder.cs
+++ b/src/Grpc.AspNetCore.Server/Internal/GrpcServiceBinder.cs
@@ -21,6 +21,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Grpc.Core;
+using Grpc.Core.Interceptors;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;