@JamesNK 2019-04-27
I think the service configuration is fine, but a more idiomatic ASP.NET way of defining filters would be placing attributes on the service implementation class. Should investigate what that would look like.