@JunTaoLuo 2019-05-14
Added a microbenchmark and here are the results:
Before:
```
          Method |     Mean |     Error |    StdDev |      Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
---------------- |---------:|----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
 HandleCallAsync | 1.713 us | 0.0341 us | 0.0365 us | 583,885.6 |      0.0439 |           - |           - |               280 B |
```
After:
```
          Method |     Mean |     Error |    StdDev |      Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
---------------- |---------:|----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
 HandleCallAsync | 1.766 us | 0.0335 us | 0.0329 us | 566,280.2 |      0.0496 |           - |           - |               320 B |

                     Method |     Mean |     Error |    StdDev |      Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
--------------------------- |---------:|----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
 InterceptedHandleCallAsync | 2.185 us | 0.0362 us | 0.0338 us | 457,706.6 |      0.1106 |           - |           - |               696 B |
```